+++ src/entrypoints/injectedButton.content/game-manager/utils/html.ts
@@ -7,6 +7,9 @@ function getWebsiteRowsImpl() {
 export const getWebsiteRows = memoize(getWebsiteRowsImpl);
 
 // Selectors for the fields from the tr Element for each row
+export const languageElSelector = 'td select[name^="idLanguage"]';
+export const conditionElSelector = 'td select[name^="idCondition"]';
+export const signedElSelector = 'td input[name^="isSigned"]';
+export const commentElSelector = 'td input[name^="comments"]';
 export const quantityElSelector = 'td input[name^="amount"]';
 export const priceElSelector = 'td input[name^="price"]';
-export const languageElSelector = 'td select[name^="idLanguage"]';